jmwallet/tests/test_address.py
"""Tests for Bitcoin address validation as used by the send command.

These tests ensure that:
- Valid addresses are accepted and produce the correct scriptPubKey.
- Addresses with invalid checksums (e.g. single-char typos) are rejected.
- Addresses from the wrong network are rejected.
- All four supported networks (mainnet, testnet, signet, regtest) work.
- Case-insensitive decoding works (uppercase addresses from QR codes).
"""

from __future__ import annotations

import pytest
from bitcointx import ChainParams
from bitcointx.wallet import CCoinAddress, CCoinAddressError

# ---------------------------------------------------------------------------
# Helpers
# ---------------------------------------------------------------------------

NETWORK_TO_CHAIN = {
    "mainnet": "bitcoin",
    "testnet": "bitcoin/testnet",
    "signet": "bitcoin/signet",
    "regtest": "bitcoin/regtest",
}


def address_to_scriptpubkey(network: str, address: str) -> bytes:
    """Parse and validate an address, returning its scriptPubKey."""
    with ChainParams(NETWORK_TO_CHAIN[network]):
        return bytes(CCoinAddress(address).to_scriptPubKey())


# ---------------------------------------------------------------------------
# Test vectors
# ---------------------------------------------------------------------------

# (network, address, expected_scriptpubkey_hex)
VALID_ADDRESSES = [
    # P2WPKH — mainnet
    (
        "mainnet",
        "bc1qw508d6qejxtdg4y5r3zarvary0c5xw7kv8f3t4",
        "0014751e76e8199196d454941c45d1b3a323f1433bd6",
    ),
    # P2WPKH — regtest
    (
        "regtest",
        "bcrt1qw508d6qejxtdg4y5r3zarvary0c5xw7kygt080",
        "0014751e76e8199196d454941c45d1b3a323f1433bd6",
    ),
    # P2WSH — mainnet
    (
        "mainnet",
        "bc1qrp33g0q5c5txsp9arysrx4k6zdkfs4nce4xj0gdcccefvpysxf3qccfmv3",
        "00201863143c14c5166804bd19203356da136c985678cd4d27a1b8c6329604903262",
    ),
    # P2WPKH — testnet
    (
        "testnet",
        "tb1qw508d6qejxtdg4y5r3zarvary0c5xw7kxpjzsx",
        "0014751e76e8199196d454941c45d1b3a323f1433bd6",
    ),
    # P2WPKH — signet (same tb1 prefix as testnet)
    (
        "signet",
        "tb1qw508d6qejxtdg4y5r3zarvary0c5xw7kxpjzsx",
        "0014751e76e8199196d454941c45d1b3a323f1433bd6",
    ),
    # P2WSH — signet (real address from production logs)
    (
        "signet",
        "tb1q8c44zs4rnlvjcwyuhmphp9azpvel55nv95ek4g89jnq4722sxuhsu3epez",
        None,  # just check it parses without error
    ),
]


class TestAddressValidation:
    """CCoinAddress validates checksums and rejects wrong-network addresses."""

    @pytest.mark.parametrize(
        ("network", "address", "expected_spk_hex"),
        VALID_ADDRESSES,
        ids=[
            "mainnet-p2wpkh",
            "regtest-p2wpkh",
            "mainnet-p2wsh",
            "testnet-p2wpkh",
            "signet-p2wpkh",
            "signet-p2wsh-production",
        ],
    )
    def test_valid_address_produces_scriptpubkey(
        self, network: str, address: str, expected_spk_hex: str | None
    ) -> None:
        spk = address_to_scriptpubkey(network, address)
        assert len(spk) > 0
        if expected_spk_hex is not None:
            assert spk.hex() == expected_spk_hex

    @pytest.mark.parametrize(
        ("network", "address"),
        [
            # Last character changed — checksum mismatch
            ("mainnet", "bc1qw508d6qejxtdg4y5r3zarvary0c5xw7kv8f3t5"),
            # Character changed in the middle
            ("mainnet", "bc1qw508d6qejxtdg4y5r3zarvary0c5xw7kv8f3s4"),
            # Regtest — last char changed
            ("regtest", "bcrt1qw508d6qejxtdg4y5r3zarvary0c5xw7kygt081"),
            # Signet — last char changed
            ("signet", "tb1qw508d6qejxtdg4y5r3zarvary0c5xw7kxpjzsq"),
        ],
        ids=["mainnet-last-char", "mainnet-mid-char", "regtest-last-char", "signet-last-char"],
    )
    def test_invalid_checksum_rejected(self, network: str, address: str) -> None:
        """A single-character typo must be caught — the original vulnerability.

        The hand-rolled decoder stripped the checksum without verifying it, so
        these addresses would have silently produced wrong scriptPubKeys and
        sent funds to unspendable outputs.
        """
        with pytest.raises(CCoinAddressError):
            address_to_scriptpubkey(network, address)

    @pytest.mark.parametrize(
        ("network", "address"),
        [
            # mainnet address decoded as regtest
            ("regtest", "bc1qw508d6qejxtdg4y5r3zarvary0c5xw7kv8f3t4"),
            # signet/testnet (tb1) address decoded as mainnet
            ("mainnet", "tb1qw508d6qejxtdg4y5r3zarvary0c5xw7kxpjzsx"),
            # regtest (bcrt1) address decoded as signet
            ("signet", "bcrt1qw508d6qejxtdg4y5r3zarvary0c5xw7kygt080"),
        ],
        ids=["mainnet-as-regtest", "signet-as-mainnet", "regtest-as-signet"],
    )
    def test_wrong_network_rejected(self, network: str, address: str) -> None:
        with pytest.raises(CCoinAddressError):
            address_to_scriptpubkey(network, address)

    def test_garbage_rejected(self) -> None:
        with pytest.raises(CCoinAddressError):
            address_to_scriptpubkey("mainnet", "notanaddress")

    def test_empty_string_rejected(self) -> None:
        with pytest.raises(CCoinAddressError):
            address_to_scriptpubkey("mainnet", "")

    def test_uppercase_accepted(self) -> None:
        """Uppercase addresses (e.g. from QR decoders) must be accepted."""
        spk = address_to_scriptpubkey("mainnet", "BC1QW508D6QEJXTDG4Y5R3ZARVARY0C5XW7KV8F3T4")
        assert spk.hex() == "0014751e76e8199196d454941c45d1b3a323f1433bd6"
