jmwallet/pyproject.toml
[project]
name = "jmwallet"
version = "0.18.0"
description = "JoinMarket wallet library with pluggable blockchain backends"
requires-python = ">=3.11"
dependencies = [
    # coincurve <21.0.0 does not support Python 3.14
    # coincurve 21.0.0 has a bug that prevents installation on 3.14
    # See: https://github.com/ofek/coincurve/issues/219
    # Pinned to commit hash until a fixed version is released
    "coincurve @ https://github.com/ofek/coincurve/archive/7829b29c08ebb1cc80386a1cdaf8c2243c4ef5c5.zip",
    "cryptography>=46.0.0",
    "httpx>=0.25.0",
    "loguru>=0.7.0",
    "mnemonic>=0.21",
    "pydantic>=2.0.0",
    "python-bitcointx>=1.1.5",
    "typer>=0.9.0",
]

[project.scripts]
jm-wallet = "jmwallet.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.5.0",
    "ruff>=0.1.0",
    "pdbpp",
    "memray",
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=jmwallet --cov-report=html --cov-report=term --timeout=60"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "docker: marks tests requiring any Docker services",
    "e2e: marks tests requiring Docker e2e profile",
    "reference: marks tests requiring Docker reference profile",
    "neutrino: marks tests requiring Docker neutrino profile",
    "reference_maker: marks tests requiring Docker reference-maker profile",
    "flaky: marks tests that may fail due to external dependencies",
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
